+++ src/main/java/com/add/venture/model/ParticipanteGrupoId.java
@@ -13,5 +13,5 @@ public class ParticipanteGrupoId implements Serializable {
     private static final long serialVersionUID = 1L;
     
     private Long usuario; // Cambiado a Long para coincidir con el tipo de Usuario.idUsuario
-    private Integer grupo; // Mantiene Integer para coincidir con GrupoViaje.idGrupo
+    private Long grupo; // Mantiene Integer para coincidir con GrupoViaje.idGrupo
 }
